refactor: OpenAPI 호출 코드 함수로 분리
