+++ backend/src/utils/openapi.11st.ts
@@ -1,8 +1,9 @@
 import { BASE_URL_11ST, OPEN_API_KEY_11ST } from 'src/constants';
 import * as convert from 'xml-js';
 import * as iconv from 'iconv-lite';
+import axios from 'axios';
 
-export function xmlConvert11st(xml: Buffer) {
+function xmlConvert11st(xml: Buffer) {
     const xmlUtf8 = iconv.decode(xml, 'EUC-KR').toString();
     const {
         ProductInfoResponse: { Product },
@@ -14,11 +15,25 @@ export function xmlConvert11st(xml: Buffer) {
     return Product;
 }
 
-export function productInfo11st(productCode: string) {
+function productInfoUrl11st(productCode: string) {
     const shopUrl = new URL(BASE_URL_11ST);
     shopUrl.searchParams.append('key', OPEN_API_KEY_11ST);
     shopUrl.searchParams.append('productCode', productCode);
     return shopUrl.toString();
+}
+
+export async function getProductInfo11st(productCode: string) {
+    const openApiUrl = productInfoUrl11st(productCode);
+    const xml = await axios.get(openApiUrl, { responseType: 'arraybuffer' });
+    const productDetails = xmlConvert11st(xml.data);
+    const price = productDetails['ProductPrice']['LowestPrice']['text'].replace(/(원|,)/g, '');
+    return {
+        productCode: productDetails['ProductCode']['text'],
+        productName: productDetails['ProductName']['text'],
+        productPrice: parseInt(price),
+        shop: '11번가',
+        imageUrl: productDetails['BasicImage']['text'],
+    };
 }
 
 export function createUrl11st(productCode: string) {
